+++ examiner/run_tests.py
@@ -60,7 +60,7 @@ def run_testcases(suite):
     """
     Run testsuit.
     """
-    runner = unittest.TextTestRunner(resultclass=RESULT_CLASS, verbosity=2)
+    runner = unittest.TextTestRunner(resultclass=RESULT_CLASS, verbosity=2, failfast=ARGS.failfast)
 
     try:
         results = runner.run(suite)